Fixing up some things
