+++ lib/screens/cupertino_main_scaffold.dart
@@ -1,0 +1,54 @@
+import 'package:Date_Roulette/constants.dart';
+import 'package:Date_Roulette/screens/tab_item.dart';
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+
+class CupertinoMainScaffold extends StatelessWidget {
+  const CupertinoMainScaffold({
+    Key key,
+    @required this.currentTab,
+    @required this.onSelectTab,
+    @required this.widgetBuilders,
+  }) : super(key: key);
+
+  final TabItem currentTab;
+  final ValueChanged<TabItem> onSelectTab;
+  final Map<TabItem, WidgetBuilder> widgetBuilders;
+
+  @override
+  Widget build(BuildContext context) {
+    return CupertinoTabScaffold(
+      tabBar: CupertinoTabBar(
+        items: [
+          _buildItem(TabItem.discover),
+          _buildItem(TabItem.chamber),
+          _buildItem(TabItem.profile),
+        ],
+        onTap: (index) => onSelectTab(TabItem.values[index]),
+      ),
+      tabBuilder: (context, index) {
+        final item = TabItem.values[index];
+        return CupertinoTabView(builder: (context) {
+          return CupertinoTabView(
+            builder: (context) => widgetBuilders[item](context),
+          );
+        });
+      },
+    );
+  }
+
+  BottomNavigationBarItem _buildItem(TabItem tabItem) {
+    final itemData = TabItemData.allTabs[tabItem];
+    final color = currentTab == tabItem ? kButtonPrimaryColor : kColorBlack;
+    return BottomNavigationBarItem(
+      icon: Icon(
+        itemData.icon,
+        color: color,
+      ),
+      title: Text(itemData.title,
+          style: TextStyle(
+            color: color,
+          )),
+    );
+  }
+}
